fix #107
All the wierd types were added to enforce that every key of an override could be undefined, a string, or a constructor. Previously a key would only have to be defined as one of those types to pass.

This will force override implementers to handle all 3 types in their code, as the compiler will start complaining that they are not.

Signed-off-by: Wright, Christopher R <Christopher.Wright@ca.com>